+++ features/todo/components/skin_imports.go
@@ -1,0 +1,9 @@
+// Package components registers all built-in UI skins at init time so
+// the skin dispatcher can resolve them at runtime.
+package components
+
+import (
+	_ "github.com/calionauta/gogogo-fullstack-template/web/skins/basecoat"
+	_ "github.com/calionauta/gogogo-fullstack-template/web/skins/daisyui"
+	_ "github.com/calionauta/gogogo-fullstack-template/web/skins/morpheus"
+)